refactor(ccw.core): create ccw.editors.clojure.text
This new package will contain everything related to text modification/strategies/ops.
